Refactor configuration management: move envs to src/config, implement validation, and add global exception filter
